Update to use new page level event listener gatherer
